pyproject.toml
[tool.poetry]
name = "travel-planner-bot"
version = "0.1.0"
description = "AI-powered travel planning bot for iMessage using Stagehand for web automation"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
chromadb = ">=0.5.0"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"
numpy = "^1.26.0,<2.0.0"
imessage-bot-framework = {git = "https://github.com/shrey150/imessage-bots.git"}
stagehand = "^0.5.4"
nest-asyncio = "^1.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

